SWIK-1356-AppendSlides search panel refactored from modal. New component added
